Making changes to the code to remove the guid_to_user table.  It's Yagni. To do it right, let's start from the simplest possible implementation, that is, to just have as a cookie the user id and then check if they are logged in. from there, let's design our way to a solid implementation.
How I killed the guid table:
1) kill guid_to_user creation script
2) kill register_user_cookie procedure
In Database_access, delete line 200, 254, 316, 335, 331, and 455 and their dependencies
which will cause the following cascade of work to be needed:

line 200:
* rename Security.unregister_user to logout_user(int userid) in logout.jsp and Security.java
* delete "delete_cookie" in Security.java and Database_access.java

line 254:
* delete look_for_logged_in_user_by_cookie(String cookie_value) from Database_access and its call in Security
* create a new method, user_is_logged_in which will check if the user is logged in, and rename get_user_from_cookie to check_if_logged_in(int userid)
* delete get_user_from_cookie in Security

Line 316:
* delete get_new_cookie_for_user() from Database_access and its call in Security.register_user()
register_user used to return a string of the cookie.  what happened with it?
it just used user_id, so I make it return void and just send user_id in as the cookie value.

TODO: we should also add a piece of code in logout.jsp that will delete the user's qarma cookie.