BK-#8-#144- Lots of stuff.  See below:
Making the sample data script more robust - now you can run it
multiple times without having to clean the database first.  Replace
is a cool mysql syntax which will insert, or delete if necessary and
then insert.

Adding localization for the new "change password" functionality.
Adding the new change password pages.

adding auditing for user login and logout

fixing bug in audit of checking cookie - when in dev mode, the
passphrase will constantly change and junk will appear in the audits.
This precludes that -  in decrypt_cookie_and_check_validity

Adding an audit for losing a point when creating a requestoffer.